Update test graphics module for nested static figures
